feat: grabTransaction in utils
